+++ CHANGELOG.rst
@@ -5,7 +5,7 @@ v/1.0.0 (2022)
  - Converted to python 3 only.
  - Adds `recursive_getattr()`, `recursive_setattr()`, and `recursive_hasattr()`.
  - Adds `pandas.move_column_inplace()`.
- - Adds `path.InOutPaths` for handling in-/output path.
+ - Adds `path.IOPaths` for handling in-/output paths.
  - Adds `path.mk_dir` for creating missing directories.
  - Adds `time.Timestamps` for recording timestamps during runtime.
  - Adds `time.StepTimer` for timing a step in a `with` statement/context.